backend/requirements.txt
# ─────────────────────────────────────────────────────────

# RAG PDF App — Backend Dependencies

# Install: pip install -r requirements.txt

# Python  : 3.11 recommended

# ─────────────────────────────────────────────────────────

# ── Flask core ────────────────────────────────────────────

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1

# ── Database ──────────────────────────────────────────────

SQLAlchemy==2.0.25
alembic==1.13.1

# PostgreSQL driver (uncomment only if using PostgreSQL)

# psycopg2-binary==2.9.9

# ── Cache ─────────────────────────────────────────────────

redis==5.0.1

# ── AWS S3 ────────────────────────────────────────────────

boto3==1.34.11
botocore==1.34.11

# ── LangChain + LLM ──────────────────────────────────────

langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.12.0

# ── Vector store ─────────────────────────────────────────

# Use faiss-gpu instead if you have CUDA available

faiss-cpu==1.13.0

# ── PDF processing ───────────────────────────────────────

pypdf==3.17.4

# Required by LangChain's PyPDFLoader

pdfminer.six==20221105

# ── Tokenisation (needed by LangChain/OpenAI) ────────────

tiktoken==0.5.2

# ── Google Auth ───────────────────────────────────────────

google-auth==2.27.0
requests==2.31.0

# ── Config / utils ────────────────────────────────────────

python-dotenv==1.0.1

# ── Production server ────────────────────────────────────

gunicorn==21.2.0
